finish user signup/loggin
