--- controllers/userControllers.js
@@ -1,95 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const bcrypt = require("bcrypt");
-const { User } = require("../models");
-const jwt = require("jsonwebtoken");
-
-// USER SIGN UP
-router.post("/", (req, res) => {
-  User.create({
-    username: req.body.username,
-    password: req.body.password,
-    email: req.body.email,
-  })
-    .then((newUser) => {
-      const token = jwt.sign(
-        {
-          username: newUser.username,
-          id: newUser.id,
-        },
-        process.env.JWT_SECRET,
-        {
-          expiresIn: "3h",
-        }
-      );
-      res.json({
-        token,
-        user: newUser,
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.json({ msg: "Error In Creating New User", err });
-    });
-});
-// USER LOGGIN ROUTE
-router.post("/login", (req, res) => {
-  //1. FIND THE CORRECT USER ROUTE
-  User.findOne({
-    where: {
-      email: req.body.email,
-    },
-  })
-    .then((foundUser) => {
-      //Wrong email catch
-      if (!foundUser) {
-        return res.status(401).json({ msg: "Email & Password Does Not Match" });
-      }
-      //Wrong Email?
-      if (!bcrypt.compareSync(req.body.password, foundUser.password)) {
-        return res.status(401).json({ msg: "Email & Password Does Not Match" });
-      }
-      //Email and password are correct at this point
-      const token = jwt.sign(
-        {
-          username: foundUser.username,
-          id: foundUser.id,
-        },
-        process.env.JWT_SECRET,
-        {
-          expiresIn: "3h",
-        }
-      );
-      res.json({
-        token,
-        user: foundUser,
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.json({ msg: "Error User Loggin", err });
-    });
-});
-
-router.get("/isValidToken", (req, res) => {
-  const token = req.headers?.authorization?.split(" ")[1];
-  if (!token) {
-    return res
-      .status(403)
-      .json({ isValid: false, msg: "Must Be Loggin!" });
-  }
-  try {
-    const tokenData = jwt.verify(token,process.env.JWT_SECRET);
-    res.json({
-      isValid: true,
-      user: tokenData,
-    });
-  } catch (err) {
-    res.status(403).json({
-      isValid: false,
-      msg: "Invalid!",
-    });
-  }
-});
-
-module.exports = router;
